candidate-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai==1.40.0
pinecone-client==3.0.0
pydantic==2.9.0
httpx==0.27.0
python-multipart==0.0.9
PyPDF2==3.0.1
pdfplumber==0.11.0

